refactor: 상품 현황 테이블 헤드 함수 추출해라
