+++ src/app/asset/components/Row/InventoryHeadRow.js
@@ -1,0 +1,15 @@
+import Row from './index.js';
+import Cell from '../Cell/index.js';
+import TABLE_TITLE from '../../constants/TABLE_TITLE.js';
+
+const InventoryHeadRow = () => {
+    const ret = Row();
+
+    ret.append(Cell(TABLE_TITLE.productName));
+    ret.append(Cell(TABLE_TITLE.productPrice));
+    ret.append(Cell(TABLE_TITLE.productQuantity));
+
+    return ret;
+};
+
+export default InventoryHeadRow;
